feat(db): run the product loop on community PostgreSQL (#1154)
Audited ontology class and relation writes commit the definition and its
chained audit decision in one transaction on PostgreSQL. RuntimeDb leaves the
async worker before synchronous PostgreSQL calls, so gRPC handlers no longer
panic. Namespace-role lookup and credential creation no longer fail closed;
policy-scoped listing and budget usage bind integer parameters at the inferred
width; principal-scoped listing hides reserved governance kinds. A
spawned-binary conformance test drives the loop on both backends (#1086).

+++ src/db/postgres_budget.rs
@@ -410,7 +410,7 @@ impl PostgresDb {
                 .execute(
                     "INSERT INTO chisei_budget_usage
                         (scope_id, metric, period_start, amount_used)
-                     VALUES ($1, $2, $3, GREATEST($4, 0))
+                     VALUES ($1, $2, $3, GREATEST($4::bigint, 0))
                      ON CONFLICT(scope_id, metric, period_start) DO UPDATE SET
                         amount_used = GREATEST(chisei_budget_usage.amount_used + $4, 0)",
                     &[&scope, &metric, &period_start, &delta],
@@ -420,7 +420,7 @@ impl PostgresDb {
                 .execute(
                     "INSERT INTO chisei_budget_attributions
                         (source_scope_id, applied_scope_id, metric, period_start, amount_used)
-                     VALUES ($1, $2, $3, $4, GREATEST($5, 0))
+                     VALUES ($1, $2, $3, $4, GREATEST($5::bigint, 0))
                      ON CONFLICT (source_scope_id, applied_scope_id, metric, period_start)
                      DO UPDATE SET
                         amount_used = GREATEST(
@@ -511,7 +511,7 @@ impl PostgresDb {
                 .execute(
                     "INSERT INTO chisei_budget_usage
                         (scope_id, metric, period_start, amount_used)
-                     VALUES ($1, $2, $3, GREATEST($4, 0))
+                     VALUES ($1, $2, $3, GREATEST($4::bigint, 0))
                      ON CONFLICT(scope_id, metric, period_start) DO UPDATE SET
                         amount_used = GREATEST(chisei_budget_usage.amount_used + $4, 0)",
                     &[&scope, &metric, &period_start, &amount],
@@ -521,7 +521,7 @@ impl PostgresDb {
                 .execute(
                     "INSERT INTO chisei_budget_attributions
                         (source_scope_id, applied_scope_id, metric, period_start, amount_used)
-                     VALUES ($1, $2, $3, $4, GREATEST($5, 0))
+                     VALUES ($1, $2, $3, $4, GREATEST($5::bigint, 0))
                      ON CONFLICT (source_scope_id, applied_scope_id, metric, period_start)
                      DO UPDATE SET
                         amount_used = GREATEST(
